+++ packages/api/src/Domain/Client/Workspace/WorkspaceApiOperations.ts
@@ -1,4 +1,7 @@
 export enum WorkspaceApiOperations {
   Creating,
   Inviting,
+  Accepting,
+  ListingWorkspaces,
+  ListingWorkspaceUsers,
 }