+++ packages/api/src/Domain/Client/Workspace/WorkspaceApiService.spec.ts
@@ -1,6 +1,10 @@
-import { WorkspaceType } from '@standardnotes/common'
-import { WorkspaceInvitationResponse } from '../../Response'
+import { WorkspaceAccessLevel, WorkspaceType } from '@standardnotes/common'
+import { HttpStatusCode } from '../../Http'
 import { WorkspaceCreationResponse } from '../../Response/Workspace/WorkspaceCreationResponse'
+import { WorkspaceInvitationAcceptingResponse } from '../../Response/Workspace/WorkspaceInvitationAcceptingResponse'
+import { WorkspaceInvitationResponse } from '../../Response/Workspace/WorkspaceInvitationResponse'
+import { WorkspaceListResponse } from '../../Response/Workspace/WorkspaceListResponse'
+import { WorkspaceUserListResponse } from '../../Response/Workspace/WorkspaceUserListResponse'
 import { WorkspaceServerInterface } from '../../Server/Workspace/WorkspaceServerInterface'
 
 import { WorkspaceApiOperations } from './WorkspaceApiOperations'
@@ -19,6 +23,17 @@ describe('WorkspaceApiService', () => {
     workspaceServer.inviteToWorkspace = jest.fn().mockReturnValue({
       data: { uuid: 'i-1-2-3' },
     } as jest.Mocked<WorkspaceInvitationResponse>)
+    workspaceServer.acceptInvite = jest.fn().mockReturnValue({
+      data: { success: true },
+    } as jest.Mocked<WorkspaceInvitationAcceptingResponse>)
+    workspaceServer.listWorkspaces = jest.fn().mockReturnValue({
+      status: HttpStatusCode.Success,
+      data: { ownedWorkspaces: [], joinedWorkspaces: [] },
+    } as jest.Mocked<WorkspaceListResponse>)
+    workspaceServer.listWorkspaceUsers = jest.fn().mockReturnValue({
+      status: HttpStatusCode.Success,
+      data: { users: [] },
+    } as jest.Mocked<WorkspaceUserListResponse>)
   })
 
   it('should create a workspace', async () => {
@@ -87,6 +102,7 @@ describe('WorkspaceApiService', () => {
     const response = await createService().inviteToWorkspace({
       workspaceUuid: 'w-1-2-3',
       inviteeEmail: 'test@test.te',
+      accessLevel: WorkspaceAccessLevel.WriteAndRead,
     })
 
     expect(response).toEqual({
@@ -97,6 +113,7 @@ describe('WorkspaceApiService', () => {
     expect(workspaceServer.inviteToWorkspace).toHaveBeenCalledWith({
       workspaceUuid: 'w-1-2-3',
       inviteeEmail: 'test@test.te',
+      accessLevel: 'write-and-read',
     })
   })
 
@@ -111,6 +128,7 @@ describe('WorkspaceApiService', () => {
       await service.inviteToWorkspace({
         workspaceUuid: 'w-1-2-3',
         inviteeEmail: 'test@test.te',
+        accessLevel: WorkspaceAccessLevel.WriteAndRead,
       })
     } catch (caughtError) {
       error = caughtError
@@ -129,7 +147,157 @@ describe('WorkspaceApiService', () => {
       await createService().inviteToWorkspace({
         workspaceUuid: 'w-1-2-3',
         inviteeEmail: 'test@test.te',
+        accessLevel: WorkspaceAccessLevel.WriteAndRead,
       })
+    } catch (caughtError) {
+      error = caughtError
+    }
+
+    expect(error).not.toBeNull()
+  })
+
+  it('should accept invite to a workspace', async () => {
+    const response = await createService().acceptInvite({
+      userUuid: 'u-1-2-3',
+      inviteUuid: 'i-1-2-3',
+      publicKey: 'foo',
+      encryptedPrivateKey: 'bar',
+    })
+
+    expect(response).toEqual({
+      data: {
+        success: true,
+      },
+    })
+    expect(workspaceServer.acceptInvite).toHaveBeenCalledWith({
+      userUuid: 'u-1-2-3',
+      inviteUuid: 'i-1-2-3',
+      publicKey: 'foo',
+      encryptedPrivateKey: 'bar',
+    })
+  })
+
+  it('should not accept invite to a workspace if it is already accepting', async () => {
+    const service = createService()
+    Object.defineProperty(service, 'operationsInProgress', {
+      get: () => new Map([[WorkspaceApiOperations.Accepting, true]]),
+    })
+
+    let error = null
+    try {
+      await service.acceptInvite({
+        userUuid: 'u-1-2-3',
+        inviteUuid: 'i-1-2-3',
+        publicKey: 'foo',
+        encryptedPrivateKey: 'bar',
+      })
+    } catch (caughtError) {
+      error = caughtError
+    }
+
+    expect(error).not.toBeNull()
+  })
+
+  it('should not accept invite to a workspace if the server fails', async () => {
+    workspaceServer.acceptInvite = jest.fn().mockImplementation(() => {
+      throw new Error('Oops')
+    })
+
+    let error = null
+    try {
+      await createService().acceptInvite({
+        userUuid: 'u-1-2-3',
+        inviteUuid: 'i-1-2-3',
+        publicKey: 'foo',
+        encryptedPrivateKey: 'bar',
+      })
+    } catch (caughtError) {
+      error = caughtError
+    }
+
+    expect(error).not.toBeNull()
+  })
+
+  it('should list workspaces', async () => {
+    const response = await createService().listWorkspaces()
+
+    expect(response).toEqual({
+      status: 200,
+      data: {
+        ownedWorkspaces: [],
+        joinedWorkspaces: [],
+      },
+    })
+    expect(workspaceServer.listWorkspaces).toHaveBeenCalled()
+  })
+
+  it('should not list workspaces if it is already listing them', async () => {
+    const service = createService()
+    Object.defineProperty(service, 'operationsInProgress', {
+      get: () => new Map([[WorkspaceApiOperations.ListingWorkspaces, true]]),
+    })
+
+    let error = null
+    try {
+      await service.listWorkspaces()
+    } catch (caughtError) {
+      error = caughtError
+    }
+
+    expect(error).not.toBeNull()
+  })
+
+  it('should not list workspaces if the server fails', async () => {
+    workspaceServer.listWorkspaces = jest.fn().mockImplementation(() => {
+      throw new Error('Oops')
+    })
+
+    let error = null
+    try {
+      await createService().listWorkspaces()
+    } catch (caughtError) {
+      error = caughtError
+    }
+
+    expect(error).not.toBeNull()
+  })
+
+  it('should list workspace users', async () => {
+    const response = await createService().listWorkspaceUsers({ workspaceUuid: 'w-1-2-3' })
+
+    expect(response).toEqual({
+      status: 200,
+      data: {
+        users: [],
+      },
+    })
+    expect(workspaceServer.listWorkspaceUsers).toHaveBeenCalledWith({ workspaceUuid: 'w-1-2-3' })
+  })
+
+  it('should not list workspace users if it is already listing them', async () => {
+    const service = createService()
+    Object.defineProperty(service, 'operationsInProgress', {
+      get: () => new Map([[WorkspaceApiOperations.ListingWorkspaceUsers, true]]),
+    })
+
+    let error = null
+    try {
+      await service.listWorkspaceUsers({ workspaceUuid: 'w-1-2-3' })
+    } catch (caughtError) {
+      error = caughtError
+    }
+
+    expect(error).not.toBeNull()
+  })
+
+  it('should not list workspace users if the server fails', async () => {
+    workspaceServer.listWorkspaceUsers = jest.fn().mockImplementation(() => {
+      throw new Error('Oops')
+    })
+
+    let error = null
+    try {
+      await createService().listWorkspaceUsers({ workspaceUuid: 'w-1-2-3' })
     } catch (caughtError) {
       error = caughtError
     }