+++ packages/api/src/Domain/Request/Workspace/WorkspaceInvitationRequestParams.ts
@@ -1,7 +1,8 @@
-import { Uuid } from '@standardnotes/common'
+import { Uuid, WorkspaceAccessLevel } from '@standardnotes/common'
 
 export type WorkspaceInvitationRequestParams = {
   workspaceUuid: Uuid
   inviteeEmail: string
+  accessLevel: WorkspaceAccessLevel
   [additionalParam: string]: unknown
 }