+++ packages/api/src/Domain/Request/Workspace/WorkspaceUserListRequestParams.ts
@@ -1,0 +1,6 @@
+import { Uuid } from '@standardnotes/common'
+
+export type WorkspaceUserListRequestParams = {
+  workspaceUuid: Uuid
+  [additionalParam: string]: unknown
+}