+++ packages/api/src/Domain/Response/Workspace/WorkspaceUserListResponseBody.ts
@@ -1,0 +1,5 @@
+import { WorkspaceUser } from '@standardnotes/models'
+
+export type WorkspaceUserListResponseBody = {
+  users: Array<WorkspaceUser>
+}