+++ packages/api/src/Domain/Server/Workspace/WorkspaceServer.spec.ts
@@ -1,7 +1,10 @@
-import { WorkspaceType } from '@standardnotes/common'
-import { HttpServiceInterface } from '../../Http'
+import { WorkspaceAccessLevel, WorkspaceType } from '@standardnotes/common'
+import { HttpServiceInterface, HttpStatusCode } from '../../Http'
 import { WorkspaceCreationResponse } from '../../Response/Workspace/WorkspaceCreationResponse'
+import { WorkspaceInvitationAcceptingResponse } from '../../Response/Workspace/WorkspaceInvitationAcceptingResponse'
 import { WorkspaceInvitationResponse } from '../../Response/Workspace/WorkspaceInvitationResponse'
+import { WorkspaceListResponse } from '../../Response/Workspace/WorkspaceListResponse'
+import { WorkspaceUserListResponse } from '../../Response/Workspace/WorkspaceUserListResponse'
 
 import { WorkspaceServer } from './WorkspaceServer'
 
@@ -40,12 +43,60 @@ describe('WorkspaceServer', () => {
     const response = await createServer().inviteToWorkspace({
       inviteeEmail: 'test@test.te',
       workspaceUuid: 'w-1-2-3',
+      accessLevel: WorkspaceAccessLevel.WriteAndRead,
     })
 
     expect(response).toEqual({
       data: {
         uuid: 'i-1-2-3',
       },
+    })
+  })
+
+  it('should accept invitation to a workspace', async () => {
+    httpService.post = jest.fn().mockReturnValue({
+      data: { success: true },
+    } as jest.Mocked<WorkspaceInvitationAcceptingResponse>)
+
+    const response = await createServer().acceptInvite({
+      encryptedPrivateKey: 'foo',
+      inviteUuid: 'i-1-2-3',
+      publicKey: 'bar',
+      userUuid: 'u-1-2-3',
+    })
+
+    expect(response).toEqual({
+      data: {
+        success: true,
+      },
+    })
+  })
+
+  it('should list workspaces', async () => {
+    httpService.get = jest.fn().mockReturnValue({
+      status: HttpStatusCode.Success,
+      data: { ownedWorkspaces: [], joinedWorkspaces: [] },
+    } as jest.Mocked<WorkspaceListResponse>)
+
+    const response = await createServer().listWorkspaces({})
+
+    expect(response).toEqual({
+      status: 200,
+      data: { ownedWorkspaces: [], joinedWorkspaces: [] },
+    })
+  })
+
+  it('should list workspace users', async () => {
+    httpService.get = jest.fn().mockReturnValue({
+      status: HttpStatusCode.Success,
+      data: { users: [] },
+    } as jest.Mocked<WorkspaceUserListResponse>)
+
+    const response = await createServer().listWorkspaceUsers({ workspaceUuid: 'w-1-2-3' })
+
+    expect(response).toEqual({
+      status: 200,
+      data: { users: [] },
     })
   })
 })