+++ packages/services/src/Domain/Workspace/WorkspaceManager.ts
@@ -1,5 +1,6 @@
 import { WorkspaceApiServiceInterface } from '@standardnotes/api'
-import { Uuid, WorkspaceType } from '@standardnotes/common'
+import { Uuid, WorkspaceAccessLevel, WorkspaceType } from '@standardnotes/common'
+import { Workspace, WorkspaceUser } from '@standardnotes/models'
 
 import { InternalEventBusInterface } from '../Internal/InternalEventBusInterface'
 import { AbstractService } from '../Service/AbstractService'
@@ -13,7 +14,58 @@ export class WorkspaceManager extends AbstractService implements WorkspaceClient
     super(internalEventBus)
   }
 
-  async inviteToWorkspace(dto: { inviteeEmail: string; workspaceUuid: Uuid }): Promise<{ uuid: string } | null> {
+  async listWorkspaceUsers(dto: { workspaceUuid: string }): Promise<{ users: WorkspaceUser[] }> {
+    try {
+      const result = await this.workspaceApiService.listWorkspaceUsers(dto)
+
+      if (result.data.error !== undefined) {
+        return { users: [] }
+      }
+
+      return result.data
+    } catch (error) {
+      return { users: [] }
+    }
+  }
+
+  async listWorkspaces(): Promise<{ ownedWorkspaces: Workspace[]; joinedWorkspaces: Workspace[] }> {
+    try {
+      const result = await this.workspaceApiService.listWorkspaces()
+
+      if (result.data.error !== undefined) {
+        return { ownedWorkspaces: [], joinedWorkspaces: [] }
+      }
+
+      return result.data
+    } catch (error) {
+      return { ownedWorkspaces: [], joinedWorkspaces: [] }
+    }
+  }
+
+  async acceptInvite(dto: {
+    inviteUuid: string
+    userUuid: string
+    publicKey: string
+    encryptedPrivateKey: string
+  }): Promise<{ success: boolean }> {
+    try {
+      const result = await this.workspaceApiService.acceptInvite(dto)
+
+      if (result.data.error !== undefined) {
+        return { success: false }
+      }
+
+      return result.data
+    } catch (error) {
+      return { success: false }
+    }
+  }
+
+  async inviteToWorkspace(dto: {
+    inviteeEmail: string
+    workspaceUuid: Uuid
+    accessLevel: WorkspaceAccessLevel
+  }): Promise<{ uuid: string } | null> {
     try {
       const result = await this.workspaceApiService.inviteToWorkspace(dto)