+++ packages/web/src/javascripts/Components/Preferences/Panes/Appearance.tsx
@@ -59,11 +59,6 @@ const Appearance: FunctionComponent<Props> = ({ application }) => {
       })
 
     themesAsItems.unshift({
-      label: 'Dark',
-      value: 'Dark',
-    })
-
-    themesAsItems.unshift({
       label: 'Default',
       value: 'Default',
     })