+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.23.204](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.203...@standardnotes/desktop@3.23.204) (2022-10-11)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.203](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.202...@standardnotes/desktop@3.23.203) (2022-10-11)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.202](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.201...@standardnotes/desktop@3.23.202) (2022-10-10)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.23.201](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.200...@standardnotes/desktop@3.23.201) (2022-10-10)
 
 **Note:** Version bump only for package @standardnotes/desktop